IA cleanup: retire /explore, move map to /standards/strategic-plan/map
Implements ADR 0003. The strategic-plan coverage map now lives at a
dedicated URL under /standards (where the data lives), the redundant
/explore tiles view is gone, and the sidebar shrinks from six entries
to five.

Surface changes:
- New: /standards/strategic-plan/map (server component, reuses the
  existing client ProjectMapView).
- New: "Map" entry on the Standards sub-nav. Active-state logic
  refactored to longest-prefix-match so the parent "Strategic Plan"
  entry is correctly NOT active when viewing the Map child.
- Removed: app/explore/page.tsx, components/ExploreViewToggle.tsx,
  the "Explore" sidebar entry, and the now-unused "search" icon case.
- Removed: the "Explore" tile from the landing steering grid (#250
  shipped 5-tile; this drops to 4-tile, 2x2 layout).

Redirects (next.config.mjs):
- /explore?view=map → /standards/strategic-plan/map (308)
- /explore[/*] → /portfolio (308)

Code paths updated:
- ProjectMap click on a category node → /portfolio?category=<slug>
  (was /explore#category-<slug>; the by-category browse now lives on
  /portfolio's filter chips per ADR 0003).
- Conversational-agent system prompt no longer suggests /explore as a
  fallback surface; suggests /portfolio (with category filter) or the
  new map URL instead.

Docs swept: CLAUDE.md (IA table + cut-routes rule), README.md,
CONTRIBUTING.md, .impeccable.md (work-categories taxonomy paragraph
and stat-strip section). Lib comments referencing /explore updated
to point at the new home. ADR 0001 historical references left as-is
(ADRs are append-only).

closes #252
