+++ .impeccable.md
@@ -209,10 +209,10 @@ the eyebrow is fixed (e.g., *"Projects"* on every visit to
 ## Work Categories Taxonomy
 
 `lib/work-categories.ts` defines the *"by problem"* axis used on
-`/explore`, on `/portfolio` filter chips, and on each portfolio card.
-The taxonomy is audience-facing — labels are written in a Dean's
-vocabulary (*"Working with documents,"* *"Reconciling accounts and
-records"*), not technical jargon.
+`/portfolio` filter chips and on each portfolio card. The taxonomy is
+audience-facing — labels are written in a Dean's vocabulary
+(*"Working with documents,"* *"Reconciling accounts and records"*),
+not technical jargon.
 
 The canonical add / rename / retire / promote policy lives in the
 header comment of `lib/work-categories.ts`. Pointers, not duplication —
@@ -263,16 +263,13 @@ projects don't follow the IIDS operational ladder, so the primary
 stage chip carries the whole signal — a second neutral chip would imply
 a granularity that doesn't exist for them.
 
-**Stat-strip and `/explore` breakdowns.** Public stage only. The
-operational granularity isn't useful at landing-density. Format:
+**Stat-strip breakdowns.** Public stage only. The operational
+granularity isn't useful at landing-density. Format:
 
 ```
 14 projects across 10 home units
 6 building · 6 live · 2 tracked
 ```
-
-`/explore` tile breakdowns follow the same shape but suppress when the
-entire category sits in one stage (avoids `4 live` tautology).
 
 **Filters on `/portfolio`.** Two tiers: a top-level Stage row (5 chips
 in `PUBLIC_STAGE_ORDER`), and a drill-in Operational Status row that
